C#: Move configuring external editor to a separate page
The configuration of an external editor doesn't fit the C# basics page. It has also grown enough that I think it's worth moving to its own page.

Removed mentions of MonoDevelop since the editor is deprecated so it shouldn't be promoted (although it's technically still supported).

Changes to the VSCode section to include the example configuration files instead of linking to a GitHub comment, and removed mentions of Mono and the Godot C# extension (which doesn't support 4.x).

Also added a note to the general external editor documentation because some C# users keep finding it instead of the C#-specific documentation.